uploaded pictures for hw2
Color pallet was focusing on setting the hue to be in the blue family to be able to nicely separate differences between precipitation density data and avoid red-green dependence.

Color illusion was made with inspiration from gray squares on a gray W->E gradient, but instead uses a Blue->Yellow gradient and Green Squares to emphasize that you can abstract over this with any two adjacent color families and just use a square which is the combination of them both. When you travel across the gradient from cool->warm or dark->light, then the middleman looks warm/light in one side and cool/dark on the other side, making it seem like a different color!